Rename config.js to siteMetadata.js
